backend/auth_bridge/backend/pyproject.toml
[project]
name = "lostfilm-auth-bridge"
version = "0.1.0"
description = "QR and pairing auth bridge for LostFilm Android TV"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "jinja2>=3.1.0",
  "pydantic-settings>=2.4.0",
  "uvicorn[standard]>=0.30.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
